Preprocessing overhaul.
Updated the preprocess function to use less redundant temporary variables, and to build up a more coherant
metadata tree in memory.

Note that this change only concerns the preprocess function, and changes made will require modification of the
set_*_point functions and the build_dropdowns function, as well as knock-on effects throughout the rest of the
generation code.