Preprocessor changed to scan steps even when excluded.
The preprocessor will now scan steps even if it has determined that the step, module, or even theme will not be written to the
generated course. This is required as the definition of a glossary term or reference may be present in a step that will be filtered
out and, if it is not scanned anyway, the resulting course will have references to a term or reference, but no definition for it.
Interlink anchors *ARE NOT* recorded, and links to anchors that are filtered out will generate a warning as would normally happen for
a link to an unknown anchor.