Utility function and glossary/reference block fixups.
Modified the utility functions to reflect the change form storing steps by filename to id in the metadata, and the introduction of
the output_id to support filtering. Corrected the templating in the glossary/reference block generator and the invocations of the
function to use the new template engine system.