Modified set_glossary_point to remove hashref argument.
As set_glossary_point is passed $self implicitly, there is no real need to pass $self -> {"terms"} as an argument as it is simpler
just to access that directly within the function, and it does nothing but add another argument to a function that already has too many
of them. This removes the need to pass the terms hash ref explicitly.