Updated template directory handling.
Updated the way in which the template module stores the location from which it will pull templates. Instead of joining the template
base directory and theme during each template operation, the path is calculated once when created or set and cached. This also allows
other modules to use the "templatedir" value to pull data from the template directory easily.