Replaced handling of include resource filtering.
includes_filter() has been overhauled to provide the same filtering logic as filter(), except that it expects the include
and exclude data to be specified as comma separated values in attributes. This will include a resource if:

- no include or exclude are set.
- no include is set, and no filters are specified by the user (ie: exclude rules could never match)
- no exclude values match filters specified by the user, and no includes are set
- an include rule matches a filter specified by the user

If a resource hash an include set, it will only be included if it does not have an exclude set that matches a user-set
filter AND one of its includes matches a user-set filter.