Fixed up validation of theme metadata.
The includes section of the theme metadata was not being validated correctly: not only was it being validated in the wrong place
(inside the module validation loop), it was also only being validated correctly if the includes element had non-text children. In the
situation where the user inadvertently does

<include><![CDATA[ ... text... ]]></include>

it would result in an attempt to use the include as a hash ref when it was a string. The new code now checks that the includes is a
hashref, and that it contains at least one resource element.