includes_filter can handle scalar argument.
If a resource in an includes block has no include or exclude attributes, it will probably be added to the resource array as a scalar
rather than a hashref (as XML::Simple has no granularity of control for ForceContent we can't turn it on, or other things will
break, and without it there's no way to force resource elements to always have -> {"content"}). This change allows it to handle that.