Added the ability to vary the AJAX request delay in stage 3.
While planning to give the user the option to increase the verbosity of export and processing output, it became clear that the size of
the resulting logs (sometimes > 100k) could be potentially troublesome to the server if the request frequency was too high. This code
allows the delay between AJAX requests to be increased depending on the verbosity level selected.