Adding wiki username storage to logged-in data.
This change allows the username the user logged into the wiki with to be stored with the login flag and information in the session
data. This is not needed by the main web ui script, but is required by the help script to identify the user properly.

Also took the opportunity to clean out old document comments.