Identified major issue with step filtering.
This fixes a serious problem with step filtering: despite calculating the output id for steps during preprocessing, the system was
then completely ignoring the output id and using the step id directly, regardless of whether this would cause discontinuities! This
forces the use of the output id by modifying get_step_name to look for the output id for a step.