Wiki support tools now require wiki_url, api_url, and upload_url.
Previous versions of the tools fell back on APEcs wikis if the user did not specify a default. This behaviour is now removed, and
users must explicitly specify the wiki url, api url, and upload url to the wiki tools either in their ~/.courseprocessor.cfg or on the
command line.